Move getCoordinateSystem util
