+++ src/server/common/helpers/session-cache/utils.js
@@ -1,5 +1,4 @@
 import { clone } from '@hapi/hoek'
-import { COORDINATE_SYSTEMS } from '~/src/server/common/constants/exemptions.js'
 
 export const EXEMPTION_CACHE_KEY = 'exemption'
 
@@ -53,21 +52,6 @@ export const resetExemptionSiteDetails = (request) => {
   delete existingCache.siteDetails
   request.yar.set(EXEMPTION_CACHE_KEY, existingCache)
   return { siteDetails: null }
-}
-
-/**
- * @param { Request } request
- */
-export const getCoordinateSystem = (request) => {
-  const existingCache = getExemptionCache(request)
-  const currentSystem = existingCache.siteDetails?.coordinateSystem
-
-  const coordinateSystem =
-    currentSystem === COORDINATE_SYSTEMS.OSGB36
-      ? COORDINATE_SYSTEMS.OSGB36
-      : COORDINATE_SYSTEMS.WGS84
-
-  return { coordinateSystem }
 }
 
 /**
